backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# Base de datos
psycopg2-binary>=2.9,<3.0

# Autenticación JWT
djangorestframework-simplejwt>=5.3,<6.0

# Celery
celery>=5.3,<6.0
redis>=5.0,<6.0

# Procesamiento de archivos Excel
openpyxl>=3.1,<4.0
pandas>=2.1,<3.0
xlrd>=2.0,<3.0

# Utilidades
python-decouple>=3.8,<4.0
python-dateutil>=2.8,<3.0
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0

# Desarrollo
ipython>=8.0,<9.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
